refactored main to decouple linear from torch libraries
