+++ libmultilabel/linear/__init__.py
@@ -1,4 +1,4 @@
 from .linear import *
-from .metrics import get_metrics
+from .metrics import get_metrics, tabulate_metrics
 from .preprocessor import *
 from .utils import *
